requirements.txt
joblib
pandas
scikit-learn
streamlitp